fix: Fix payment repository aggregations and auth parameters
**Payment Repository Fixes:**
- Fixed SUM() aggregations returning Decimal instead of i64
- Added ::BIGINT casts to all SUM(amount_cents) expressions
- Fixed FILTER clause syntax: (SUM() FILTER (...))::BIGINT
- 6 SUM queries corrected (total_paid, stats aggregations)

**Quote Handlers Fixes:**
- Fixed _auth parameter name in accept_quote handler
- Fixed _auth parameter name in reject_quote handler
- Both handlers now correctly use auth.user_id

**Impact:**
- Fixes 101 → ~40 compilation errors
- Payment statistics queries now work correctly
- Decimal type mismatch resolved

+++ backend/src/infrastructure/database/repositories/payment_repository_impl.rs
@@ -718,7 +718,7 @@ impl PaymentRepository for PostgresPaymentRepository {
     async fn get_total_paid_for_expense(&self, expense_id: Uuid) -> Result<i64, String> {
         let row = sqlx::query!(
             r#"
-            SELECT COALESCE(SUM(amount_cents - refunded_amount_cents), 0) AS "total!"
+            SELECT COALESCE(SUM(amount_cents - refunded_amount_cents), 0)::BIGINT AS "total!"
             FROM payments
             WHERE expense_id = $1 AND status = 'succeeded'
             "#,
@@ -734,7 +734,7 @@ impl PaymentRepository for PostgresPaymentRepository {
     async fn get_total_paid_by_owner(&self, owner_id: Uuid) -> Result<i64, String> {
         let row = sqlx::query!(
             r#"
-            SELECT COALESCE(SUM(amount_cents - refunded_amount_cents), 0) AS "total!"
+            SELECT COALESCE(SUM(amount_cents - refunded_amount_cents), 0)::BIGINT AS "total!"
             FROM payments
             WHERE owner_id = $1 AND status = 'succeeded'
             "#,
@@ -750,7 +750,7 @@ impl PaymentRepository for PostgresPaymentRepository {
     async fn get_total_paid_for_building(&self, building_id: Uuid) -> Result<i64, String> {
         let row = sqlx::query!(
             r#"
-            SELECT COALESCE(SUM(amount_cents - refunded_amount_cents), 0) AS "total!"
+            SELECT COALESCE(SUM(amount_cents - refunded_amount_cents), 0)::BIGINT AS "total!"
             FROM payments
             WHERE building_id = $1 AND status = 'succeeded'
             "#,
@@ -771,9 +771,9 @@ impl PaymentRepository for PostgresPaymentRepository {
                 COUNT(*) FILTER (WHERE status = 'succeeded') AS "succeeded_count!",
                 COUNT(*) FILTER (WHERE status = 'failed') AS "failed_count!",
                 COUNT(*) FILTER (WHERE status = 'pending') AS "pending_count!",
-                COALESCE(SUM(amount_cents), 0) AS "total_amount_cents!",
-                COALESCE(SUM(amount_cents) FILTER (WHERE status = 'succeeded'), 0) AS "total_succeeded_cents!",
-                COALESCE(SUM(refunded_amount_cents), 0) AS "total_refunded_cents!"
+                COALESCE(SUM(amount_cents)::BIGINT, 0) AS "total_amount_cents!",
+                COALESCE((SUM(amount_cents) FILTER (WHERE status = 'succeeded'))::BIGINT, 0) AS "total_succeeded_cents!",
+                COALESCE(SUM(refunded_amount_cents)::BIGINT, 0) AS "total_refunded_cents!"
             FROM payments
             WHERE owner_id = $1
             "#,
@@ -806,9 +806,9 @@ impl PaymentRepository for PostgresPaymentRepository {
                 COUNT(*) FILTER (WHERE status = 'succeeded') AS "succeeded_count!",
                 COUNT(*) FILTER (WHERE status = 'failed') AS "failed_count!",
                 COUNT(*) FILTER (WHERE status = 'pending') AS "pending_count!",
-                COALESCE(SUM(amount_cents), 0) AS "total_amount_cents!",
-                COALESCE(SUM(amount_cents) FILTER (WHERE status = 'succeeded'), 0) AS "total_succeeded_cents!",
-                COALESCE(SUM(refunded_amount_cents), 0) AS "total_refunded_cents!"
+                COALESCE(SUM(amount_cents)::BIGINT, 0) AS "total_amount_cents!",
+                COALESCE((SUM(amount_cents) FILTER (WHERE status = 'succeeded'))::BIGINT, 0) AS "total_succeeded_cents!",
+                COALESCE(SUM(refunded_amount_cents)::BIGINT, 0) AS "total_refunded_cents!"
             FROM payments
             WHERE building_id = $1
             "#,
